+++ src/cpu/x64/gemm_bf16_inner_product.hpp
@@ -283,7 +283,7 @@ struct gemm_bf16_inner_product_bwd_weights_t : public primitive_t {
             dim_t OCB_per_thread = utils::div_up(OCB, bias_reduction_nthr_);
 
             OC_per_thread = OCB_per_thread * bias_blksize;
-            nthr_OCB = utils::div_up(OCB, OCB_per_thread);
+            nthr_OCB = static_cast<int>(utils::div_up(OCB, OCB_per_thread));
             nthr_MB = bias_reduction_nthr_ / nthr_OCB;
 
             assert(nthr_OCB * nthr_MB <= bias_reduction_nthr_);
